Rename paths to fix lint
